pyproject.toml
[project]
name = "conf-agg"
version = "0.1.0"
description = "Confidence Aggregation for Mathematical Reasoning"
authors = [
    {name = "Conf_Agg Team", email = "team@confagg.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Core ML Libraries
    "torch>=2.1.0",
    "transformers>=4.51.0",
    "accelerate>=0.24.0",
    "peft>=0.6.0",
    
    # Data Processing
    "datasets>=2.14.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    
    # Configuration & Logging
    "hydra-core>=1.3.0",
    "wandb>=0.15.0",
    "tqdm>=4.65.0",
    
    # Math Verification
    "math-verify>=0.1.0",
    
    # Reinforcement Learning
    "trl>=0.7.0",
    
    # Additional Dependencies
    "tokenizers>=0.15.0",
    "safetensors>=0.4.0",
    "protobuf>=4.25.0",
    
    # Ray Serve for parallel processing
    "ray[serve]>=2.8.0",
    
    # vLLM for inference
    "vllm>=0.2.0",
]

[project.optional-dependencies]
dev = [
    # Development Tools
    "jupyter>=1.0.0",
    "ipykernel>=6.25.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "jupyter>=1.0.0",
    "ipykernel>=6.25.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88
